Added a GraphQLResponse return object for GraphQLTestUtils perform
This will allow for future updates to that response type, e.g. in line
with (ResultActions)[https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/test/web/servlet/ResultActions.html]. This will allow more fluid checks of the json
response instead of having to manually traverse the `JsonNode` tree.
The only reason for this return object at this stage is to prepare the
public API for those additional features in the future.